matrix load complete

+++ src/matrix.cpp
@@ -3,6 +3,7 @@
 //
 #include <sstream>
 #include <algorithm>
+#include <cassert>
 #include "matrix.h"
 #include "global.h"
 
@@ -10,6 +11,7 @@ Matrix::Matrix(const string &matName) {
     logger.log("Matrix::Matrix");
     this->sourceFileName = "../data/" + matName + ".csv";
     this->matrixName = matName;
+    this->maxDimPerBlock = floor(sqrt((1024 / sizeof(int)) * BLOCK_SIZE));
 }
 
 
@@ -31,29 +33,123 @@ bool Matrix::extractSize(const string &firstline) {
     string word;
     stringstream s(firstline);
     while (getline(s, word, ',')) {
-        word.erase(std::remove_if(word.begin(), word.end(), ::isspace), word.end());
         if (word.empty()) {
             return false;
         }
     }
     auto cnt = count(firstline.begin(), firstline.end(), ',');
-    this->size = cnt;
+    if (firstline.back() == ',') cnt--;
+    this->size = cnt + 1;
+    this->blockCount = (this->size + this->maxDimPerBlock - 1) / this->maxDimPerBlock;
     return true;
 }
 
 bool Matrix::blockify() {
     logger.log("Matrix::blockify");
     fstream fin(this->sourceFileName, ios::in);
+    // decide whether sparse or not
+    size_t zero{};
+    string line;
+    for (size_t i = 0; i < this->size; i++) {
+        getline(fin, line);
+        std::stringstream ss(line);
+        std::string cell;
+        for (size_t j = 0; j < this->size; j++) {
+            if (!getline(ss, cell, ','))
+                return false;
+            auto p = stoi(cell);
+            if (p == 0LL) zero++;
+        }
+    }
+    if ((double) zero >= 0.6 * (double) this->size * (double) this->size) this->isSparse = true;
+    fin.close();
+    auto cnt = this->size;
 
+
+    while (cnt >= this->maxDimPerBlock) {
+        cnt -= this->maxDimPerBlock;
+        this->columnsPerBlockCount.push_back(this->maxDimPerBlock);
+    }
+    if (cnt > 0) this->columnsPerBlockCount.push_back(cnt);
+    assert(this->columnsPerBlockCount.size() == this->blockCount);
+    this->rowsPerBlockCount = this->columnsPerBlockCount;
+    if (this->isSparse) {
+
+    } else {
+        for (size_t j = 0; j < this->blockCount; j++) {
+            ifstream filein(this->sourceFileName, ios::in);
+            for (size_t i = 0; i < this->blockCount; i++) {
+                vector<vector<int>> data(this->rowsPerBlockCount[i]);
+                for (size_t row = 0; row < this->rowsPerBlockCount[i]; row++) {
+                    getline(filein, line);
+                    stringstream ss(line);
+                    string word;
+                    size_t col;
+                    // seek to appropriate column number
+                    for (col = 0; col < j * this->maxDimPerBlock; col++) {
+                        if (!getline(ss, word, ',')) {
+                            return false;
+                        }
+                    }
+                    vector<int> rowdata;
+                    for (col = 0; col < this->columnsPerBlockCount[j]; col++) {
+                        if (!getline(ss, word, ',')) {
+                            return false;
+                        }
+                        rowdata.push_back(stoi(word));
+                    }
+                    data[row] = rowdata;
+                }
+                BufferManager::writePage(this->matrixName, i, j, data, this->rowsPerBlockCount[i],
+                                         this->columnsPerBlockCount[j]);
+            }
+        }
+        return true;
+    }
     return false;
 }
 
 void Matrix::transpose() {
+    if (this->isSparse) {
 
+    } else {
+        for (size_t r = 0; r < this->blockCount; r++) {
+            for (size_t c = r + 1; c < this->blockCount; c++) {
+                auto p1 = bufferManager.getPage(this->matrixName, r, c);
+                auto p2 = bufferManager.getPage(this->matrixName, c, r);
+                for (size_t i = 0; i < this->rowsPerBlockCount[r]; i++) {
+                    for (size_t j = 0; j < this->columnsPerBlockCount[c]; j++) {
+                        swap(p1.data[i][j], p2.data[j][i]);
+                    }
+                }
+                BufferManager::writePage(this->matrixName, r, c, p2.data, this->rowsPerBlockCount[r],
+                                         this->columnsPerBlockCount[c]);
+                BufferManager::writePage(this->matrixName, c, r, p1.data, this->rowsPerBlockCount[c],
+                                         this->columnsPerBlockCount[r]);
+                bufferManager.updatePage(this->matrixName, r, c);
+                bufferManager.updatePage(this->matrixName, c, r);
+            }
+        }
+        for (size_t r = 0; r < this->blockCount; r++) {
+            auto p1 = bufferManager.getPage(this->matrixName, r, r);
+            for (size_t i = 0; i < this->rowsPerBlockCount[r]; i++) {
+                for (size_t j = i + 1; j < this->columnsPerBlockCount[r]; j++) {
+                    swap(p1.data[i][j], p1.data[j][i]);
+                }
+            }
+            BufferManager::writePage(this->matrixName, r, r, p1.data, this->rowsPerBlockCount[r],
+                                     this->columnsPerBlockCount[r]);
+            bufferManager.updatePage(this->matrixName, r, r);
+        }
+    }
 }
 
 void Matrix::print() const {
+    if (this->isSparse) {
 
+    } else {
+
+    }
 }
 
 void Matrix::makePermanent() {
@@ -62,8 +158,13 @@ void Matrix::makePermanent() {
         BufferManager::deleteFile(this->sourceFileName);
     string newSourceFile = "../data/" + this->matrixName + ".csv";
     ofstream fout(newSourceFile, ios::out);
+    if (this->isSparse) {
 
+    } else {
+        for (size_t r = 0; r < this->blockCount; r++) {
 
+        }
+    }
 }
 
 bool Matrix::isPermanent() const {
@@ -75,9 +176,13 @@ bool Matrix::isPermanent() const {
 
 void Matrix::unload() const {
     logger.log("Matrix::unload");
-    for (long long int i = 0; i < this->blockCount; i++) {
-        for (long long int j = 0; j < this->blockCount; j++) {
-            bufferManager.deleteFile(this->matrixName, i, j);
+    if (this->isSparse) {
+
+    } else {
+        for (size_t i = 0; i < this->blockCount; i++) {
+            for (size_t j = 0; j < this->blockCount; j++) {
+                bufferManager.deleteFile(this->matrixName, i, j);
+            }
         }
     }
 }
